User Privilege with orders creation

+++ routes/api.php
@@ -23,17 +23,27 @@ Route::middleware(['auth:api', 'admin'])->group(function () {
 
     // Frames CRUD
 
-    Route::post('frames', [App\Http\Controllers\FramesController::class, 'store']);
-    Route::put('frames/{id}', [App\Http\Controllers\FramesController::class, 'update']);
-    Route::delete('frames/{id}', [App\Http\Controllers\FramesController::class, 'destroy']);
+    Route::post('frames', [App\Http\Controllers\API\Admin\FramesController::class, 'store']);
+    Route::put('frames/{id}', [App\Http\Controllers\API\Admin\FramesController::class, 'update']);
+    Route::delete('frames/{id}', [App\Http\Controllers\API\Admin\FramesController::class, 'destroy']);
 
     //  Lenses CRUD
 
-    Route::post('lens', [App\Http\Controllers\LensesController::class, 'store']);
-    Route::put('lens/{id}', [App\Http\Controllers\LensesController::class, 'update']);
-    Route::delete('lens/{id}', [App\Http\Controllers\LensesController::class, 'destroy']);
+    Route::post('lens', [App\Http\Controllers\API\Admin\LensesController::class, 'store']);
+    Route::put('lens/{id}', [App\Http\Controllers\API\Admin\LensesController::class, 'update']);
+    Route::delete('lens/{id}', [App\Http\Controllers\API\Admin\LensesController::class, 'destroy']);
 
 });
+
+Route::middleware('auth:api')->group(function () {
+    Route::get('frames/active', [App\Http\Controllers\API\User\FramesController::class, 'getActiveFrames']);
+
+    Route::get('lense/active', [App\Http\Controllers\API\User\LensesController::class, 'getActiveLenses']);
+
+    Route::get('orders/create', [App\Http\Controllers\API\User\OrdersController::class, 'createCustomGlasses']);
+
+});
+
 
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
